refactor: 광고 플랫폼 판별·그룹화 유틸 분리
